0.4.0: Use read instead of read_until.
Read in up to 8K chunks; if many lines come in at once it's misleading
to bother timing them.  If a line fragment comes in, write it out as
soon as possible.

Code is a bit uglier, but it also produces tidier and less misleading
output and works better with programs that output partial lines.