Experimental pty support - issue #4
Need to clean it up, gate it to systems with openpty(), and put it
behind an optional flag.

With this I can run tcpdump under rtss and get proper real-time output
instead of it coming in giant spurts.

top output is predictably mangled.  Maybe a flag to scrub output of
cursor-movement commands?